+++ Source/TerritoryFramework/Private/Core/TerritoryDeveloperSettings.cpp
@@ -2,7 +2,6 @@
 
 UTerritoryDeveloperSettings::UTerritoryDeveloperSettings()
 {
-	DefaultPlayerFaction = FGameplayTag::RequestGameplayTag(FName(TEXT("Narrative.Factions.Heroes")), false);
 	DefaultNarrativeButtonClass = TSoftClassPtr<UNarrativeCommonButtonBase>(FSoftObjectPath(
 		TEXT("/NarrativePro/Pro/Core/UI/Widgets/Base/WBP_NarrativeButton_Text.WBP_NarrativeButton_Text_C")));
 	DefaultTerritoryTextStyle = TSoftClassPtr<UCommonTextStyle>(FSoftObjectPath(